requirements.txt
cython
keras=2.1.5 
opencv 
tensorflow 
numpy >=1.14
matplotlib 
pytables 
pandas 
pyqt
scipy 
scikit-learn 
scikit-image
statsmodels 
numba
tqdm
requests
